Refactoring

+++ skyler/src/main/resources/com/skyler/smarthome/server/controller/gwcontroller/GatewayInternalPushController.java
@@ -1,5 +1,6 @@
 package com.skyler.smarthome.server.controller.gwcontroller;
 
+import com.skyler.smarthome.server.data.DeviceDao;
 import com.skyler.smarthome.server.data.GatewayDao;
 import com.skyler.smarthome.server.data.ModuleDao;
 import com.skyler.smarthome.server.model.Gateway;
@@ -30,63 +31,69 @@ public class GatewayInternalPushController {
 	GatewayDao gatewayDAO;
 	@Autowired(required=true)
 	ModuleDao moduleDao;
+	@Autowired(required=true)
+	DeviceDao deviceDao;
 
 	//Register new GateWay
-	@RequestMapping(value = "/gateways/", method = RequestMethod.POST,produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
+	@RequestMapping(value = "/gateway/", method = RequestMethod.POST,produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
 	public @ResponseBody boolean newGateway() {
-		//TODO
+		//gatewayDAO.createGateway();
 		return true;
 	}
 
 	//Delete GateWay
-	@RequestMapping(value = "/gateways/{gateway_id}", method = RequestMethod.DELETE)
+	@RequestMapping(value = "/gateway/{gateway_id}", method = RequestMethod.DELETE)
 	public void deleteGateway(@PathVariable int gateway_id) {
-		boolean result = gatewayDAO.deleteGateway(gateway_id);
+		if(gateway_id > 0) {
+			boolean result = gatewayDAO.deleteGateway(gateway_id);
+		}
 	}
 
 	//Update GateWay info
-	@RequestMapping(value = "/gateways/{gateway_id}", method = RequestMethod.PATCH)
+	@RequestMapping(value = "/gateway/{gateway_id}", method = RequestMethod.PATCH)
 	public void updateGateway(@PathVariable int gateway_id, @RequestParam String gatewayField,
 			@RequestParam String newParam) {
-		boolean result = gatewayDAO.updateGatewayByField(gateway_id, gatewayField, newParam);
+		if (gateway_id > 0&& gatewayField != null && newParam != null) {
+			boolean result = gatewayDAO.updateGatewayByField(gateway_id, gatewayField, newParam);
+		}
 	}
 
 	//Register sensor event
-	@RequestMapping(value = "/gateways/{gateway_id}/sensor/{sensor_id}/events/", method = RequestMethod.POST,produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
+	@RequestMapping(value = "/gateway/{gateway_id}/sensor/{sensor_id}/events/", method = RequestMethod.POST,produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
 	public @ResponseBody boolean registerSensorEvent(@PathVariable int gateway_id, @RequestParam int sensor_id) {
 		//TODO
 		return true;
 	}
 
 	//Register actuator event
-	@RequestMapping(value = "/gateways/{gateway_id}/actuator/{actuator_id}/events/", method = RequestMethod.POST,produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
+	@RequestMapping(value = "/gateway/{gateway_id}/actuator/{actuator_id}/events/", method = RequestMethod.POST,produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
 	public @ResponseBody boolean registerActuatorEvent(@PathVariable int gateway_id, @RequestParam int actuator_id) {
 		//TODO
 		return true;
 	}
 
 	//Register new sensor
-	@RequestMapping(value = "/gateways/{gateway_id}/sensor/", method = RequestMethod.POST,produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
+	@RequestMapping(value = "/gateway/{gateway_id}/sensor/", method = RequestMethod.POST,produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
 	public @ResponseBody boolean registerNewSensor(@PathVariable int gateway_id) {
-		//TODO
+		//moduleDao.addModuleToDevice(gateway_id,);
 		return true;
 	}
 	//Register new actuator
-	@RequestMapping(value = "/gateways/{gateway_id}/actuator/", method = RequestMethod.POST,produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
+	@RequestMapping(value = "/gateway/{gateway_id}/actuator/", method = RequestMethod.POST,produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
 	public @ResponseBody boolean registerNewActuator(@PathVariable int gateway_id) {
 		//TODO
 		return true;
 	}
 
 	//Update sensor
-	@RequestMapping(value = "/update/sensor/{sensor_id}", method = RequestMethod.POST)
+	@RequestMapping(value = "/update/sensor/{sensor_id}", method = RequestMethod.PATCH)
 	public @ResponseBody boolean updateSensor(@PathVariable int sensor_id) {
 		//TODO
 		return true;
 	}
 
 	//Update actuator
-	@RequestMapping(value = "/update/actuator/{actuator_id}", method = RequestMethod.POST)
+	@RequestMapping(value = "/update/actuator/{actuator_id}", method = RequestMethod.PATCH)
 	public @ResponseBody boolean updateActuator(@PathVariable int sensor_id) {
 		//TODO
 		return true;
